Resolution of http://code.google.com/p/jsonmarshaller/issues/detail?id=8
The SortedSet case was simply not handled. Adding a collection type enum to which the production of a fresh instance is delegated.

Fixing another bug where polymorphic entities within a collection were not marshalled polymorphically. The fix in place is more a temporary patch before the issue is investigated further. It seems like creating a descriptor of a polymorphic entity twice does not work properly.

git-svn-id: http://jsonmarshaller.googlecode.com/svn/trunk@14 30b5078c-5c40-0410-b1d4-cb505aa573e9